+++ src/master/master.hpp
@@ -1827,7 +1827,7 @@ private:
 
   double _outstanding_offers()
   {
-    return offers.size();
+    return static_cast<double>(offers.size());
   }
 
   double _event_queue_messages()
